MVP-476 prozzie setup.sh does not work in centos 7
CentOS 7 comes with a little old bash version (4.2). Alias variables
comes in bash 4.3 (`local/declare -l`), and allows to pass a variable
(or array) reference as a parameter function.

Previous version of bash does not allow to use array as parameter, so we
have to use a workaround: We use textual representation of array (as
returned in `declare -p`), and we re-create array inside function. Since
it does not modify the original array, we use a similar trick to return
a modified array: we return the textual representation, and let caller
function to create/override it.

If we drop support for bash<4.3, we can revert this commit.