WIZZIE-310 Fix linux_setup.sh execution over a new prozzie installation
After last change, execution of linux_setup.sh script in an empty
installation directory results in warnings messages.

The warnings are due to env variables treatment: they are stored in a
temporary file to be able to perform a clean exit if the user press
Ctrol-C or something bad happens, and that temp file was not created.