WIZZIE-332 Integrate monitor SNMP request in prozzie
Integration of monitor capabilities into prozzie.

This commit needs to add a volume to provide user a way to inject custom
mibs to monitor container, so monitor_setup.sh will ask for a path to
mount inside. If no path is provided, a empty volume will be mounted in
the cointainer.

On the other hand, all SNMP agents and requests needs to be provided in
an array. If no agent is provided, monitor will wait for nothing
forever.